Fix HTTP request corruption by the user agent (GH-32)
The file `VERSION` used to send the version contains a new line.

When it's used to generate the UserAgent for curl, the new line char
stops the headers sent for the request. All other headers sent after the
user agent are ignored and used like the body requests.

To solve this issue, clean up the `VERSION` file content on load.